apps/competitor-analysis-agent/pyproject.toml
[project]
name = "competitor-analysis-agent"
version = "0.1.0"
description = "ADK agent system for comprehensive competitor analysis"
requires-python = ">=3.10,<3.13"
authors = [
    {name = "ShipKit", email = "brandon@shipkit.ai"},
]
dependencies = [
    # Core ADK dependencies
    "google-adk>=1.14.1",
    "google-genai>=1.24.0",
    # Service account management
    "google-cloud-iam>=2.19.1", # For service account and key management
    "google-cloud-resource-manager>=1.14.2", # For IAM policy binding
    # Core Python dependencies
    "pydantic>=2.11.0",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.1.0",
    # Web search and data collection
    "requests>=2.32.0",
    "aiohttp>=3.9.0",
    "beautifulsoup4>=4.12.0",
    "selenium>=4.16.0",
    # Data processing
    "pandas>=2.1.0",
    "numpy>=1.25.0",
    # Database connectivity for session service
    "psycopg2-binary>=2.9.0", # PostgreSQL adapter
    "sqlalchemy>=2.0.0", # Database ORM for session management
    # FastAPI server dependencies
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6", # For form data handling
    "sseclient-py>=1.8.0",
    "google-cloud-aiplatform[adk,agent_engines]>=1.95.1", # Includes vertexai functionality
]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=25.1.0",
    "ruff>=0.12.0",
    "mypy>=1.16.0",
    "python-dotenv[cli]>=1.1.1",
]
test = [
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "responses>=0.24.0",
]

[tool.ruff]
line-length = 88
target-version = "py310"
src = ["competitor_analysis_agent"]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
    "N",  # pep8-naming
    "B904", # Exception chaining
]
ignore = [
    "E203", # whitespace before : (conflicts with black)
    "E501", # line too long (handled by black)
    "B008", # FastAPI Depends() calls in argument defaults are correct usage
]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build" 
